REFACTOR: Decomposed file bridge fiel
