+++ app/api/chat/route.ts
@@ -9,12 +9,10 @@ import { createOpenAI } from "@ai-sdk/openai";
 import { streamText, stepCountIs, type ModelMessage, tool } from "ai";
 import { getMcpTools } from "@/lib/mcp/get-mcp-tools";
 import { DEFAULT_MODEL } from "@/models";
-import {
-  downloadAttachmentsToTemp,
-  persistModifiedFiles,
-  cleanupTempDir,
-  type FileBridgeResult,
-} from "@/lib/mcp/file-bridge";
+import { downloadAttachmentsToTemp } from "@/lib/mcp/download-attachments-to-temp";
+import { persistModifiedFiles } from "@/lib/mcp/persist-modified-files";
+import { cleanupTempDir } from "@/lib/mcp/cleanup-temp-dir";
+import type { FileBridgeResult } from "@/types/file-bridge-result";
 
 type TextPart = { type: "text"; text: string };
 type ImagePart = { type: "image"; image: URL | string; mimeType?: string };
