+++ types/file-bridge-result.ts
@@ -1,0 +1,9 @@
+import { BridgedFile } from "./bridged-file";
+
+/**
+ * Result of downloading attachments to temporary staging directory.
+ */
+export type FileBridgeResult = {
+  tempDir: string;
+  files: BridgedFile[];
+};
